+++ src/functionalTest/resources/scenarios/configuration_cache/src/main/java/test/TestClassWithOptional.java
@@ -1,0 +1,15 @@
+/*
+ * Sample header 2023.
+ *
+ * File: TestClassWithOptional.java
+ *
+ * Optional sample line.
+ */
+
+package test;
+
+public class TestClassWithOptional {
+	public static void main(String[] args) {
+		System.out.println("Hello world.");
+	}
+}