+++ src/main/java/dev/yumi/gradle/licenser/api/rule/LicenseYearSelectionMode.java
@@ -8,9 +8,7 @@
 
 package dev.yumi.gradle.licenser.api.rule;
 
-import dev.yumi.gradle.licenser.YumiLicenserGradleExtension;
 import dev.yumi.gradle.licenser.util.GitUtils;
-import org.gradle.api.Project;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 
@@ -27,18 +25,18 @@ import java.util.List;
  * Represents the mode in which the year should be fetched.
  *
  * @author LambdAurora
- * @version 1.0.0
+ * @version 2.0.0
  * @since 1.0.0
  */
 public enum LicenseYearSelectionMode {
 	/**
 	 * The license year is project-wide, a change in any file of the project will update every file.
 	 */
-	PROJECT((project, path) -> project.getRootProject().getProjectDir().toPath()),
+	PROJECT((rootPath, path) -> rootPath),
 	/**
 	 * Each file has its own year.
 	 */
-	FILE((project, path) -> path);
+	FILE((rootDir, path) -> path);
 
 	private static final List<LicenseYearSelectionMode> VALUES = List.of(values());
 	private final CommitPathReference commitPathReference;
@@ -52,18 +50,17 @@ public enum LicenseYearSelectionMode {
 	 * <p>
 	 * In the case of {@link #PROJECT} the last creation year isn't file dependent.
 	 *
-	 * @param project the project the file is in
+	 * @param rootPath the root directory of the project the path is in
+	 * @param projectCreationYear the creation year of the project
 	 * @param path the path to the file
 	 * @return the creation year
 	 * @throws IOException if the file attributes cannot be read
 	 */
-	public int getCreationYear(@NotNull Project project, @NotNull Path path) throws IOException {
-		Path commitPath = this.commitPathReference.getPathForCommitFetching(project, path);
+	public int getCreationYear(@NotNull Path rootPath, int projectCreationYear, @NotNull Path path) throws IOException {
+		Path commitPath = this.commitPathReference.getPathForCommitFetching(rootPath, path);
 
 		if (commitPath != path) {
-			return project.getExtensions().getByType(YumiLicenserGradleExtension.class)
-					.getProjectCreationYear()
-					.get();
+			return projectCreationYear;
 		} else {
 			Instant instant = Files.readAttributes(commitPath, BasicFileAttributes.class).creationTime().toInstant();
 			LocalDate localDate = LocalDate.ofInstant(instant, ZoneId.systemDefault());
@@ -77,13 +74,13 @@ public enum LicenseYearSelectionMode {
 	 * <p>
 	 * In the case of {@link #PROJECT} the last modification year isn't file dependent.
 	 *
-	 * @param project the project the file is in
+	 * @param rootPath the root directory of the project the path is in
 	 * @param path the path to the file
 	 * @return the last modification year
 	 */
-	public int getModificationYear(@NotNull Project project, @NotNull Path path) {
-		Path commitPath = this.commitPathReference.getPathForCommitFetching(project, path);
-		return GitUtils.getModificationYear(project, commitPath);
+	public int getModificationYear(@NotNull Path rootPath, @NotNull Path path) {
+		Path commitPath = this.commitPathReference.getPathForCommitFetching(rootPath, path);
+		return GitUtils.getModificationYear(rootPath, commitPath);
 	}
 
 	/**
@@ -106,10 +103,10 @@ public enum LicenseYearSelectionMode {
 		/**
 		 * Gets the path to use to fetch the latest commit.
 		 *
-		 * @param project the project the path is in
+		 * @param rootPath the root directory of the project the path is in
 		 * @param path the path
 		 * @return the path to use to get the latest commit
 		 */
-		@NotNull Path getPathForCommitFetching(Project project, Path path);
+		@NotNull Path getPathForCommitFetching(Path rootPath, Path path);
 	}
 }