+++ src/main/java/dev/yumi/gradle/licenser/impl/LicenseHeader.java
@@ -10,28 +10,30 @@ package dev.yumi.gradle.licenser.impl;
 
 import dev.yumi.gradle.licenser.YumiLicenserGradlePlugin;
 import dev.yumi.gradle.licenser.api.rule.HeaderFileContext;
+import dev.yumi.gradle.licenser.api.rule.HeaderLine;
 import dev.yumi.gradle.licenser.api.rule.HeaderRule;
-import org.gradle.api.Project;
+import dev.yumi.gradle.licenser.api.rule.LicenseYearSelectionMode;
+import dev.yumi.gradle.licenser.api.rule.token.RuleToken;
+import dev.yumi.gradle.licenser.api.rule.token.TextToken;
+import dev.yumi.gradle.licenser.api.rule.token.VarToken;
+import dev.yumi.gradle.licenser.api.rule.variable.VariableType;
 import org.gradle.api.logging.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.IOException;
+import java.io.*;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 /**
  * Represents the valid license headers for this project.
  *
  * @author LambdAurora
- * @version 1.0.0
+ * @version 2.0.0
  * @since 1.0.0
  */
-public final class LicenseHeader {
-	private final List<HeaderRule> rules;
+public final class LicenseHeader implements Serializable {
+	private List<HeaderRule> rules;
 
 	public LicenseHeader(HeaderRule... rules) {
 		this(new ArrayList<>(List.of(rules)));
@@ -81,13 +83,14 @@ public final class LicenseHeader {
 	/**
 	 * Formats the given file to contain the correct license header.
 	 *
-	 * @param project the project the file is in
+	 * @param rootPath the root directory of the project the path is in
+	 * @param projectCreationYear the creation year of the project
 	 * @param logger the logger
 	 * @param path the path of the file
 	 * @param readComment the read header comment if successful, or {@code null} otherwise
 	 * @return {@code true} if files changed, otherwise {@code false}
 	 */
-	public @Nullable List<String> format(Project project, Logger logger, Path path, @Nullable List<String> readComment) {
+	public @Nullable List<String> format(Path rootPath, int projectCreationYear, Logger logger, Path path, @Nullable List<String> readComment) {
 		List<String> newHeader = null;
 
 		if (readComment == null) {
@@ -96,7 +99,7 @@ public final class LicenseHeader {
 			}
 
 			newHeader = this.format(
-					project, path, this.rules.get(0),
+					rootPath, projectCreationYear, path, this.rules.get(0),
 					new HeaderRule.ParsedData(Map.of(), Collections.emptySet(), null)
 			);
 		} else {
@@ -110,7 +113,7 @@ public final class LicenseHeader {
 						logger.lifecycle("  => Found rule in lookup.", path);
 					}
 
-					newHeader = this.format(project, path, rule, data);
+					newHeader = this.format(rootPath, projectCreationYear, path, rule, data);
 					break;
 				}
 
@@ -124,7 +127,7 @@ public final class LicenseHeader {
 					logger.lifecycle("  => Could not find rule in lookup. Using default rule.", path);
 				}
 
-				newHeader = this.format(project, path, this.rules.get(0), first);
+				newHeader = this.format(rootPath, projectCreationYear, path, this.rules.get(0), first);
 			}
 		}
 
@@ -135,14 +138,74 @@ public final class LicenseHeader {
 		}
 	}
 
-	private List<String> format(Project project, Path path, HeaderRule rule, HeaderRule.ParsedData parsed) {
+	private List<String> format(Path rootPath, int projectCreationYear, Path path, HeaderRule rule, HeaderRule.ParsedData parsed) {
 		try {
-			int creationYear = rule.getYearSelectionMode().getCreationYear(project, path);
-			int lastModifiedYear = rule.getYearSelectionMode().getModificationYear(project, path);
+			int creationYear = rule.getYearSelectionMode().getCreationYear(rootPath, projectCreationYear, path);
+			int lastModifiedYear = rule.getYearSelectionMode().getModificationYear(rootPath, path);
 			var context = new HeaderFileContext(path.getFileName().toString(), creationYear, lastModifiedYear);
 			return rule.apply(parsed, context);
 		} catch (IOException e) {
 			throw new RuntimeException(e);
 		}
+	}
+
+	@Serial
+	private void writeObject(ObjectOutputStream out) throws IOException {
+		out.writeInt(this.rules.size());
+		for (var rule : this.rules) {
+			this.writeRule(out, rule);
+		}
+	}
+
+	@Serial
+	private void readObject(ObjectInputStream in) throws IOException, ClassNotFoundException {
+		int ruleCount = in.readInt();
+		this.rules = new ArrayList<>(ruleCount);
+		for (int i = 0; i < ruleCount; i++) {
+			this.rules.add(this.readRule(in));
+		}
+	}
+
+	private void writeRule(ObjectOutput out, HeaderRule rule) throws IOException {
+		out.writeUTF(rule.getName());
+
+		out.writeObject(rule.getLines());
+
+		var variables = rule.getVariables();
+		out.writeInt(variables.size());
+		for (var entry : variables.entrySet()) {
+			out.writeUTF(entry.getKey());
+
+			for (var knownEntry : VariableType.TYPES.entrySet()) {
+				if (knownEntry.getValue() == entry.getValue()) {
+					out.writeUTF(knownEntry.getKey());
+				}
+			}
+		}
+
+		out.writeObject(rule.getYearSelectionMode());
+	}
+
+	@SuppressWarnings("unchecked")
+	private HeaderRule readRule(ObjectInput in) throws IOException, ClassNotFoundException {
+		String name = in.readUTF();
+
+		var lines = (List<HeaderLine>) in.readObject();
+
+		int variableCount = in.readInt();
+		var variables = new HashMap<String, VariableType<?>>(variableCount);
+		for (int i = 0; i < variableCount; i++) {
+			String variableName = in.readUTF();
+			String variableType = in.readUTF();
+
+			variables.put(variableName, VariableType.TYPES.get(variableType));
+		}
+
+		return new HeaderRule(
+				name,
+				lines,
+				variables,
+				(LicenseYearSelectionMode) in.readObject()
+		);
 	}
 }