+++ src/main/java/dev/yumi/gradle/licenser/task/SourceConsumer.java
@@ -9,7 +9,6 @@
 package dev.yumi.gradle.licenser.task;
 
 import dev.yumi.gradle.licenser.api.comment.HeaderComment;
-import org.gradle.api.Project;
 import org.gradle.api.logging.Logger;
 import org.jetbrains.annotations.ApiStatus;
 
@@ -20,7 +19,7 @@ import java.nio.file.Path;
  * Represents a source consumer to execute an action on project files.
  *
  * @author LambdAurora
- * @version 1.0.0
+ * @version 2.0.0
  * @since 1.0.0
  */
 @ApiStatus.Internal
@@ -28,14 +27,24 @@ interface SourceConsumer {
 	/**
 	 * Called with a project file to process.
 	 *
-	 * @param project the project
+	 * @param rootDir the root directory
+	 * @param projectCreationYear the project's creation year
+	 * @param buildPath the build directory path
 	 * @param logger the logger
-	 * @param sourceSetPath the path to the processed source set
+	 * @param projectPath the path to the project
 	 * @param headerComment the header comment type used for this file type
 	 * @param path the path to the file to process
 	 * @throws IOException if read or writing the file fails
 	 */
-	void consume(Project project, Logger logger, Path sourceSetPath, HeaderComment headerComment, Path path) throws IOException;
+	void consume(
+			Path rootDir,
+			int projectCreationYear,
+			Path buildPath,
+			Logger logger,
+			Path projectPath,
+			HeaderComment headerComment,
+			Path path
+	) throws IOException;
 
 	/**
 	 * Called when the task has finished processing all matching files.